Revert "land: Disable dex2oat watchdog"
This was initially thought to have fixed getting stucked in a black
screen after the boot animation. However the real cause was later found
to be mm-qcamera-daemon not being able to finish initializing in time
due to libc timeouts (which was also fixed). Remove the placebo fix, we
do have enough resources for dex2oat watchdog.

This reverts commit c50b91569a480cd31317d25fe6e1a0ee724c146e.

Change-Id: Ibd5cb831432cf8bbcece9a5c1695e7037859205a